+++ package.json
@@ -68,7 +68,7 @@
     "framer-motion": "^4.1.17",
     "graphql": "^15.7.2",
     "graphql-let": "^0.18.6",
-    "graphql-upload": "^12.0.0",
+    "graphql-upload": "^13.0.0",
     "i18n": "^0.13.3",
     "i18next": "^21.5.0",
     "image-size": "^1.0.0",
