+++ shared/validation/index.ts
@@ -1,15 +1,10 @@
-export { default as UserExists } from './dbValidation/IsUsedUsername';
-export { default as FollowsUser } from './dbValidation/FollowsUser';
-export { default as ExistsUser } from './dbValidation/ExistsUser';
-export { default as IsUsedUsername } from './dbValidation/IsUsedUsername';
-
-export { default as ValidateUsername } from './inputValidation/ValidateUsername';
-export { default as ValidateEmail } from './inputValidation/ValidateEmail';
-export { default as ValidatePassword } from './inputValidation/ValidatePassword';
-export { default as ValidateName } from './inputValidation/ValidateName';
+export { default as ValidateComment } from './inputValidation/ValidateComment';
 export { default as ValidateCoordinates } from './inputValidation/ValidateCoordinates';
 export { default as ValidateDate } from './inputValidation/ValidateDate';
 export { default as ValidateDescription } from './inputValidation/ValidateDescription';
-export { default as ValidateVerificationToken } from './inputValidation/ValidateVerificationToken';
+export { default as ValidateEmail } from './inputValidation/ValidateEmail';
 export { default as ValidateHashtag } from './inputValidation/ValidateHashtag';
-export { default as ValidateComment } from './inputValidation/ValidateComment';
+export { default as ValidateName } from './inputValidation/ValidateName';
+export { default as ValidatePassword } from './inputValidation/ValidatePassword';
+export { default as ValidateUsername } from './inputValidation/ValidateUsername';
+export { default as ValidateVerificationToken } from './inputValidation/ValidateVerificationToken';
