+++ src/queries/FilterPlaces.ts
@@ -1,51 +1,34 @@
-import { ValidateCoordinates, ValidateDate } from '../../shared/validation';
 import RunCypherQuery from '../database/RunCypherQuery';
 
 type queryInput = {
-  maxLatitude: number;
-  minLatitude: number;
-  maxLongitude: number;
-  minLongitude: number;
-  minDate?: number;
-  maxDate?: number;
+  filter: {
+    maxLatitude: number | null;
+    minLatitude: number | null;
+    maxLongitude: number | null;
+    minLongitude: number | null;
+    minDate: number | null;
+    maxDate: number | null;
+    radius: {
+      latitude: number;
+      longitude: number;
+      distance: number;
+    } | null;
+    tags: string[] | null;
+    skip: number | null;
+    take: number | null;
+  } | null;
+  loggedId: number | null;
 };
 
-const FilterPlacesQuery = async ({
-  maxLatitude,
-  minLatitude,
-  maxLongitude,
-  minLongitude,
-  minDate,
-  maxDate,
-}: queryInput) => {
-  // validate coordinates
-  const validateMaxCoordinates = ValidateCoordinates([
-    maxLatitude,
-    maxLongitude,
-  ]).error;
-  if (validateMaxCoordinates) throw new Error(validateMaxCoordinates);
-  const validateMinCoordinates = ValidateCoordinates([
-    minLatitude,
-    minLongitude,
-  ]).error;
-  if (validateMinCoordinates) throw new Error(validateMinCoordinates);
+const PlacesQuery = async ({ filter, loggedId }: queryInput) => {
+  const query = `
+  MATCH (place:Place)
+  WHERE (place.location.latitude >= $minLatitude OR $minLatitude IS NULL)                 // min latitude
+    AND (place.location.latitude <= $maxLatitude OR $maxLatitude IS NULL)                 // max latitude
+    AND (place.location.longitude >= $minLongitude OR $minLongitude IS NULL)              // min longitude
+    AND (place.location.longitude <= $maxLongitude OR $maxLongitude IS NULL)              // max longitude
+    AND (distance(place.location, point($radius)) <= $radius.distance OR $radius IS NULL) // radius 
 
-  // validate date if defined
-  if (minDate) {
-    const validateMinDate = ValidateDate(Number(minDate)).error;
-    if (validateMinDate) throw new Error(validateMinDate);
-  }
-  if (maxDate) {
-    const validateMaxDate = ValidateDate(Number(maxDate)).error;
-    if (validateMaxDate) throw new Error(validateMaxDate);
-  }
-
-  const query = `MATCH (place:Place)
-  WHERE place.location.latitude >= $minLatitude   // min latitude
-  AND place.location.latitude <= $maxLatitude  // max latitude
-  AND place.location.longitude >= $minLongitude    // min longitude
-  AND place.location.longitude <= $maxLongitude     // max longitude
-  
   // return most liked post from place which has photos
   CALL {
       WITH place
@@ -79,6 +62,13 @@ const FilterPlacesQuery = async ({
       icon: place.icon,
       postCount,
       likeCount,
+      distance:  
+          CASE
+              // if radius in input is not defined return null
+              WHEN $radius IS NULL THEN NULL
+              // when searching in radius return distance from center
+              ELSE distance(place.location, point($radius)) 
+          END,
       preview: 
           CASE
               // if place has preview return place preview
@@ -86,21 +76,28 @@ const FilterPlacesQuery = async ({
               // else if place has post with photos, return photos as preview otherwise return null
               ELSE post.url
           END
-  }) AS places`;
+  })[$skip..$skip + $take] // limit array 
+  AS places
+  `;
 
   const [result] = await RunCypherQuery({
     query,
     params: {
-      minLatitude,
-      maxLatitude,
-      minLongitude,
-      maxLongitude,
-      minDate,
-      maxDate,
+      // if parameter is undefined set it to null
+      minLatitude: filter?.minLatitude ?? null,
+      maxLatitude: filter?.maxLatitude ?? null,
+      minLongitude: filter?.minLongitude ?? null,
+      maxLongitude: filter?.maxLongitude ?? null,
+      minDate: filter?.minDate ?? null,
+      maxDate: filter?.maxDate ?? null,
+      skip: filter?.skip ?? 0,
+      take: filter?.take ?? 5000,
+      radius: filter?.radius ?? null,
+      loggedId,
     },
   });
 
   return result.records[0].get('places');
 };
 
-export default FilterPlacesQuery;
+export default PlacesQuery;
